add new test module for adapters
